Conflicts:
	src/main/java/com/odysseusinc/arachne/datanode/service/client/atlas/AtlasClient.java

+++ src/main/java/com/odysseusinc/arachne/datanode/service/impl/DataSourceServiceImpl.java
@@ -26,6 +26,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.base.Preconditions;
 import com.odysseusinc.arachne.commons.api.v1.dto.CommonHealthStatus;
+import com.odysseusinc.arachne.datanode.exception.IllegalOperationException;
 import com.odysseusinc.arachne.datanode.exception.NotExistException;
 import com.odysseusinc.arachne.datanode.model.datanode.DataNode;
 import com.odysseusinc.arachne.datanode.model.datasource.DataSource;
@@ -36,16 +37,17 @@ import com.odysseusinc.arachne.datanode.service.DataSourceService;
 import java.util.HashMap;
 import java.util.Map;
 import javax.annotation.PostConstruct;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
+import com.odysseusinc.arachne.execution_engine_common.api.v1.dto.DBMSType;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Transactional
@@ -114,6 +116,12 @@ public class DataSourceServiceImpl implements DataSourceService {
     public void delete(Long id) {
 
         checkNotNull(id, "given data source surrogate id is blank ");
+        final DataSource dataSource = getById(id);
+        if (dataSource.getRegistred()) {
+            final String message
+                    = String.format("Can not delete registered DataSource with id='%s'. Unregister it first", id);
+            throw new IllegalOperationException(message);
+        }
         dataSourceRepository.delete(id);
     }
 
@@ -125,12 +133,14 @@ public class DataSourceServiceImpl implements DataSourceService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Optional<DataSource> findBySid(String sid) {
 
         Preconditions.checkArgument(StringUtils.isNotBlank(sid), "given data source surrogate sid is blank ");
         return dataSourceRepository.findByUuid(sid);
     }
-
+    @Override
+    @Transactional(readOnly = true)
     public Optional<DataSource> findByCentralId(Long centralId) {
 
         Preconditions.checkArgument(Objects.nonNull(centralId), "given data source centralId is null");
@@ -147,32 +157,69 @@ public class DataSourceServiceImpl implements DataSourceService {
 
 
     @Override
-    public Optional<DataSource> update(User user, DataSource ds) {
+    @Transactional(rollbackFor = Exception.class)
+    public DataSource update(User user, DataSource dataSource) {
 
-        DataSource forUpdate = dataSourceRepository.findOne(ds.getId());
-        forUpdate.setName(ds.getName());
-        forUpdate.setType(ds.getType());
-        forUpdate.setCdmSchema(ds.getCdmSchema());
-        forUpdate.setConnectionString(ds.getConnectionString());
-        forUpdate.setDescription(ds.getDescription());
-        forUpdate.setPassword(ds.getPassword());
-        forUpdate.setUsername(ds.getUsername());
+        final DataSource exists = getById(dataSource.getId());
 
-        return Optional.of(dataSourceRepository.save(forUpdate));
+        final String name = dataSource.getName();
+        if (Objects.nonNull(name)) {
+            exists.setName(name);
+        }
+        final DBMSType type = dataSource.getType();
+        if (Objects.nonNull(type)) {
+            exists.setType(type);
+        }
+        final String cdmSchema = dataSource.getCdmSchema();
+        if (Objects.nonNull(cdmSchema)) {
+            exists.setCdmSchema(cdmSchema);
+        }
+        final String connectionString = dataSource.getConnectionString();
+        if (Objects.nonNull(connectionString)) {
+            exists.setConnectionString(connectionString);
+        }
+        final String description = dataSource.getDescription();
+        if (Objects.nonNull(description)) {
+            exists.setDescription(description);
+        }
+        final String password = dataSource.getPassword();
+        if (Objects.nonNull(password)) {
+            exists.setPassword(password);
+        }
+        final String username = dataSource.getUsername();
+        if (Objects.nonNull(username)) {
+            exists.setUsername(username);
+        }
+        final String atlasResultDbSchema = dataSource.getResultSchema();
+        if (Objects.nonNull(atlasResultDbSchema)) {
+            exists.setResultSchema(atlasResultDbSchema);
+        }
+        final String atlasTargetDbSchema = dataSource.getTargetSchema();
+        if (Objects.nonNull(atlasTargetDbSchema)) {
+            exists.setTargetSchema(atlasTargetDbSchema);
+        }
+        final String atlasTargetCohortTable = dataSource.getCohortTargetTable();
+        if (Objects.nonNull(atlasTargetCohortTable)) {
+            exists.setCohortTargetTable(atlasTargetCohortTable);
+        }
+        return dataSourceRepository.save(exists);
     }
 
     @Transactional
     @Override
-    public DataSource markDataSourceAsRegistered(Long centralId) {
+    public DataSource markDataSourceAsRegistered(DataSource dataSource, Long centralId) {
 
-        return setDSRegistered(centralId, true);
+        dataSource.setCentralId(centralId);
+        return setDSRegistered(dataSource, centralId);
     }
 
     @Transactional
     @Override
     public DataSource markDataSourceAsUnregistered(Long centralId) {
 
-        return setDSRegistered(centralId, false);
+        DataSource dataSource = dataSourceRepository.findByCentralId(centralId)
+                .orElseThrow(() -> new NotExistException(DataSource.class));
+        return setDSRegistered(dataSource, null);
     }
 
     @Transactional(rollbackFor = Exception.class)
@@ -186,13 +233,10 @@ public class DataSourceServiceImpl implements DataSourceService {
         });
     }
 
-    private DataSource setDSRegistered(Long centralId, boolean registered) {
+    private DataSource setDSRegistered(DataSource dataSource, Long centralId) {
 
-        DataSource forUpdate = dataSourceRepository.findByCentralId(centralId)
-                .orElseThrow(() -> new NotExistException(DataSource.class));
-        forUpdate.setRegistred(registered);
-        dataSourceRepository.save(forUpdate);
-        return forUpdate;
+        dataSource.setRegistred(centralId != null);
+        return dataSourceRepository.save(dataSource);
     }
 
     protected final Sort getSort(String sortBy, Boolean sortAsc) {
